requirements.txt
requests
python-dotenv
python-socketio
python-engineio
flask
flask-socketio
eventlet
itsdangerous
plyer
